sample_projects/ecommerce_api/requirements.txt
Django>=5.2.0
djangorestframework>=3.15.0
django-filter>=24.0
coreapi>=2.3.0